chore(server): move the lib folder below the app folder
