+++ server/app/controllers/validations/suggest_query.rb
@@ -51,7 +51,7 @@ module Validations
     end
 
     def valid_address?
-      Utils::CkbUtils.parse_address(query_key) rescue nil
+      CkbUtils.parse_address(query_key) rescue nil
     end
   end
 end
