+++ server/app/lib/api/v1/exceptions.rb
