+++ server/app/lib/fast_jsonapi/error_serializer.rb
